requirements.txt
nonebot2[aiohttp]>=2.0.0
nonebot-adapter-onebot>=2.2.0
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=0.12.0
python-dotenv>=0.19.0